+++ include/envoy/json/json_object.h
@@ -136,7 +136,7 @@ public:
 
   /**
    * Validates JSON object against passed in schema.
-   * @param schema supplies the string format of the schema. A Json::Exception will be thrown, if
+   * @param schema supplies the schema in string format. A Json::Exception will be thrown, if
    *        the JSON object doesn't conform to the supplied schema.
    */
   virtual void validateSchema(const std::string& schema) const PURE;
