+++ source/common/filter/ratelimit.cc
@@ -1,6 +1,7 @@
 #include "ratelimit.h"
 
 #include "common/common/empty_string.h"
+#include "common/json/config_schemas.h"
 
 namespace RateLimit {
 namespace TcpFilter {
@@ -8,6 +9,8 @@ namespace TcpFilter {
 Config::Config(const Json::Object& config, Stats::Store& stats_store, Runtime::Loader& runtime)
     : domain_(config.getString("domain")),
       stats_(generateStats(config.getString("stat_prefix"), stats_store)), runtime_(runtime) {
+
+  config.validateSchema(Json::Schema::RATELIMIT_SCHEMA);
 
   for (const Json::ObjectPtr& descriptor : config.getObjectArray("descriptors")) {
     Descriptor new_descriptor;
