--- source/common/json/schema/config_schemas.h
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace Json {
-
-class Schema {
-public:
-  static const std::string LISTENER_SCHEMA;
-};
-
-} // Json
