+++ source/server/config/network/client_ssl_auth.h
@@ -14,9 +14,6 @@ public:
   NetworkFilterFactoryCb tryCreateFilterFactory(NetworkFilterType type, const std::string& name,
                                                 const Json::Object& json_config,
                                                 Server::Instance& server);
-
-private:
-  static const std::string CLIENT_SSL_SCHEMA;
 };
 
 } // Configuration
