+++ source/server/config/network/redis_proxy.cc
@@ -7,25 +7,12 @@
 namespace Server {
 namespace Configuration {
 
-const std::string RedisProxyFilterConfigFactory::REDIS_PROXY_SCHEMA(R"EOF(
-  {
-      "$schema": "http://json-schema.org/schema#",
-      "properties":{
-        "cluster_name" : {"type" : "string"}
-      },
-      "required": ["cluster_name"],
-      "additionalProperties": false
-  }
-  )EOF");
-
 NetworkFilterFactoryCb RedisProxyFilterConfigFactory::tryCreateFilterFactory(
     NetworkFilterType type, const std::string& name, const Json::Object& config,
     Server::Instance& server) {
   if (type != NetworkFilterType::Read || name != "redis_proxy") {
     return nullptr;
   }
-
-  config.validateSchema(REDIS_PROXY_SCHEMA);
 
   Redis::ProxyFilterConfig filter_config(config, server.clusterManager());
   std::shared_ptr<Redis::ConnPool::Instance> conn_pool(new Redis::ConnPool::InstanceImpl(
