+++ source/server/config/network/tcp_proxy.h
@@ -14,9 +14,6 @@ public:
   NetworkFilterFactoryCb tryCreateFilterFactory(NetworkFilterType type, const std::string& name,
                                                 const Json::Object& config,
                                                 Server::Instance& server);
-
-private:
-  static const std::string TCP_PROXY_SCHEMA;
 };
 
 } // Configuration
