+++ source/server/configuration_impl.cc
@@ -6,6 +6,7 @@
 #include "envoy/ssl/context_manager.h"
 
 #include "common/common/utility.h"
+#include "common/json/config_schemas.h"
 #include "common/ratelimit/ratelimit_impl.h"
 #include "common/ssl/context_config_impl.h"
 #include "common/tracing/http_tracer_impl.h"
@@ -114,65 +115,13 @@ void MainImpl::initializeTracers(const Json::Object& tracing_configuration) {
 }
 
 const std::list<Server::Configuration::ListenerPtr>& MainImpl::listeners() { return listeners_; }
-const std::string MainImpl::ListenerConfig::LISTENER_SCHEMA(R"EOF(
-  {
-    "$schema": "http://json-schema.org/schema#",
-    "definitions": {
-      "ssl_context" : {
-        "type" : "object",
-        "properties" : {
-          "cert_chain_file" : {"type" : "string"},
-          "private_key_file": {"type" : "string"},
-          "alpn_protocols" : {"type" : "string"},
-          "alt_alpn_protocols": {"type" : "string"},
-          "ca_cert_file" : {"type" : "string"},
-          "verify_certificate_hash" : {"type" : "string"},
-          "verify_subject_alt_name" : {"type" : "string"},
-          "cipher_suites" : {"type" : "string"}
-        },
-        "required": ["cert_chain_file", "private_key_file"],
-        "additionalProperties": false
-      },
-      "filters" : {
-        "type" : "object",
-        "properties" : {
-          "type": {"type" : "string", "enum" :["read", "write", "both"]},
-          "name" : {
-            "type": "string",
-            "enum" : ["client_ssl_auth", "echo", "mongo_proxy", "ratelimit", "tcp_proxy", "http_connection_manager"]
-          },
-          "config": {"type" : "object"}
-        },
-        "required": ["type", "name", "config"],
-        "additionalProperties": false
-      }
-    },
-    "properties": {
-       "port": {"type": "number"},
-       "filters" : {
-         "type" : "array",
-         "minItems" : 1,
-         "items": {
-           "type": "object",
-           "properties": {"$ref" : "#/definitions/filters"}
-         }
-       },
-       "ssl_context" : {"$ref" : "#/definitions/ssl_context"},
-       "bind_to_port" : {"type": "boolean"},
-       "use_proxy_proto" : {"type" : "boolean"},
-       "use_original_dst" : {"type" : "boolean"}
-    },
-    "required": ["port", "filters"],
-    "additionalProperties": false
-  }
-  )EOF");
 
 MainImpl::ListenerConfig::ListenerConfig(MainImpl& parent, Json::Object& json)
     : parent_(parent), port_(json.getInteger("port")),
       scope_(parent_.server_.stats().createScope(fmt::format("listener.{}.", port_))) {
   log().info("  port={}", port_);
 
-  json.validateSchema(LISTENER_SCHEMA);
+  json.validateSchema(Json::Schema::LISTENER_SCHEMA);
 
   if (json.hasObject("ssl_context")) {
     Ssl::ContextConfigImpl context_config(*json.getObject("ssl_context"));
