+++ test/common/filter/ratelimit_test.cc
@@ -57,6 +57,20 @@ public:
   RequestCallbacks* request_callbacks_{};
 };
 
+TEST_F(RateLimitFilterTest, BadRatelimitConfig) {
+  std::string json_string = R"EOF(
+  {
+    "stat_prefix": "my_stat_prefix",
+    "domain" : "fake_domain",
+    "descriptors": [[{ "key" : "my_key",  "value" : "my_value" }]],
+    "ip_white_list": "12"
+  }
+  )EOF";
+
+  Json::ObjectPtr json_config = Json::Factory::LoadFromString(json_string);
+  EXPECT_THROW(Config(*json_config, stats_store_, runtime_), Json::Exception);
+}
+
 TEST_F(RateLimitFilterTest, OK) {
   InSequence s;
